% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot_cdr3aa.R
\name{barplot_cdr3aa}
\alias{barplot_cdr3aa}
\title{distribution of V/D/J usage or clonetype according to length of CDR3 aa}
\usage{
barplot_cdr3aa(input_matrix, file_out, label, gradient)
}
\arguments{
\item{input_matrix}{Firstly column should be length of CDR3 aa}

\item{file_out}{the filename of output figue,should end up with pdf or png}

\item{label}{clonetype or V/D/J segment}

\item{gradient}{the filename of output figue,should end up with pdf or png}
}
\description{
It will generate a published quality figure according to
V/d/j or clonetype and length of CDR3 aa
}
\examples{
#' barplot_cdr3aa(input_matrix,'v-usage-cdr3aa-length.pdf','Variable segment','F')
}
\keyword{circle}
