% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdr3_diversity_stat.R
\name{cdr3_diversity_stat}
\alias{cdr3_diversity_stat}
\title{statistics for CDR3 diversity}
\usage{
cdr3_diversity_stat(cdr3aa, method = c("shannon", "simpson", "gini", "d50"))
}
\arguments{
\item{cdr3aa}{A vector contain all of the cdr3 sequence}

\item{method}{which diverity should be cal:shannon/simpson/gini/d50}
}
\value{
a numeric value
}
\description{
In contrast to the previously reported Shannon index and Simpson index,
the D50 value is understandable and direct.It is the calculated percentage of dominant unique clones,
 accumulative reads of which made up for 50% of the total (ranges from 0 to 50 in theory)
}
\keyword{cdr3_diversity_stat}
