% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdr3_stat_plot.R
\name{cdr3_stat_plot}
\alias{cdr3_stat_plot}
\title{visualization for 3 kind of statistics for CDR3 aa}
\usage{
cdr3_stat_plot(stat_matrix, file_out)
}
\arguments{
\item{stat_matrix}{three columns are "type","name","value",3 rows +5 rows + 5 rows}

\item{file_out}{the filename of output figue,should end up with pdf or png}
}
\description{
The outermost layer will show the top 5  CDR3 aa and their frequencies
The middle layer will be the Returns Tukey's five number summary (minimum, lower-hinge, median, upper-hinge, maximum)
THe innermost layer will be the unique/double/3 plus CDR3 aa
}
\examples{
#' cdr3_stat_plot(stat_matrix,'cdr3aa-distribution-stat.pdf')
}
\keyword{cdr3_stat_plot}
