% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_usage.R
\name{compare_usage}
\alias{compare_usage}
\title{compare the V/D/J/cdr3aa length distribution}
\usage{
compare_usage(IRseq_results, choose, group_list)
}
\arguments{
\item{IRseq_results}{A list contains all the samples' IRseq basic stat results}

\item{choose}{v_usage/d_usage/j_usage/cdr3aa_length_usage/}

\item{group_list}{A vectors,eg 0,0,0,1,1,1,which should correspond to the samples in IRseq_results}
}
\value{
3 figures and 2 data.frame(dat and stat)
}
\description{
bar plot and t.test results
}
\examples{
#' compare_usage(all_results,c(0,0,0,1,1,1))
}
\keyword{compare_usage}
