% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgs_install.R
\name{pkgs_install}
\alias{pkgs_install}
\title{For package dependecies only}
\usage{
pkgs_install(pkgs = c("FField", "RColorBrewer", "VennDiagram", "ape"))
}
\arguments{
\item{pkgs}{A list of packages}
}
\description{
So far, I just find the way to install the packages in CRAN, not bioconductor.
It will automatelly download a list of packges and load them in current R environment.
}
\examples{
#' pkgs_install(c('FField','RColorBrewer','VennDiagram','ape'))
}
\keyword{package}
