% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cdr3aa_bar.R
\name{plot_cdr3aa_bar}
\alias{plot_cdr3aa_bar}
\title{distribution of V/D/J usage or clonetype according to length of CDR3 aa}
\usage{
plot_cdr3aa_bar(input_matrix, label, file_out)
}
\arguments{
\item{input_matrix}{Firstly column should be length of CDR3 aa}

\item{label}{clonetype or V/D/J segment}

\item{file_out}{the filename of output figue,should end up with pdf or png}
}
\description{
It will generate a published quality figure according to
V/d/j or clonetype and length of CDR3 aa
}
\examples{
#' plot_cdr3aa_bar(input_matrix,,'Variable segment','v-usage-cdr3aa-length.pdf' )
}
\keyword{circle}
