% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_IR_basic.R
\name{plot_IR_basic}
\alias{plot_IR_basic}
\title{basic bstatistics for summary table from igblastn result}
\usage{
plot_IR_basic(this_result, this_sample)
}
\arguments{
\item{stat_IR_basic(IR_basic_df)}{the results from stat_IR_basic(IR_basic_df)}
}
\value{
pdf files for V/D/J/CDR3 usage  ~~~~
}
\description{
just generate some figures
}
\examples{
#' plot_IR_basic(stat_IR_basic(IR_basic_df),'test')
}
\keyword{stat_IR_basic}
