% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_matrix_t_test.R
\name{plot_matrix_t_test}
\alias{plot_matrix_t_test}
\title{plot the t test results of the matrix}
\usage{
plot_matrix_t_test(t_test_results, file_out)
}
\arguments{
\item{t_test_results}{p/t/type column are needed, others are option}

\item{file_out}{the filename of output figue,should end up with pdf or png}
}
\description{
It will generate a published quality figure according to the V/D/J usage matrix t test results
}
\examples{
#' plot_vdj_usage_ht(v_usage_matrix,'v-usage-heatmap.pdf')
}
\keyword{heatmap}
