% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_t_test.R
\name{plot_t_test}
\alias{plot_t_test}
\title{boxplot for t.test result of two group of data}
\usage{
plot_t_test(a, b, file_out)
}
\arguments{
\item{a}{value vector for first  group in numeric}

\item{b}{value vector for second group in numeric}

\item{file_out}{the filename of output figue,should end up with pdf or png}
}
\description{
boxplot and jitter plot
}
\examples{
#' plot_t_test(stat_matrix,'cdr3aa-distribution-stat.pdf')
}
\keyword{plot_t_test}
