% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_v_j_combination.R
\name{plot_v_j_combination}
\alias{plot_v_j_combination}
\title{visualization for V-J combination}
\usage{
plot_v_j_combination(vj_usage_matrix, image_type, file_out)
}
\arguments{
\item{vj_usage_matrix}{columns and rows correspond to V and J segments respectively}

\item{image_type}{cirle/bubble/3D_barplot}

\item{file_out}{the filename of output figue,should end up with pdf or png}
}
\description{
It will generate a published quality figure according to the V and J usage marix.
}
\examples{
#' plot_v_j_combination(vj_usage_matrix,'circle','v-j-circle.pdf')
}
\keyword{combination}
