% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_vdj_usage_ht.R
\name{plot_vdj_usage_ht}
\alias{plot_vdj_usage_ht}
\title{Heatmap for V/D/J usage among samples}
\usage{
plot_vdj_usage_ht(usage_matrix, file_out)
}
\arguments{
\item{usage_matrix}{cols are V/D/J, rows are samples}

\item{file_out}{the filename of output figue,should end up with pdf or png}
}
\description{
It will generate a published quality figure according to the V/D/J usage matrix
}
\examples{
#' plot_vdj_usage_ht(v_usage_matrix,'v-usage-heatmap.pdf')
}
\keyword{heatmap}
