% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_cdr3aa_saturation.R
\name{stat_cdr3aa_saturation}
\alias{stat_cdr3aa_saturation}
\title{saturation analyses for CDR3 sequences}
\usage{
stat_cdr3aa_saturation(IRsample)
}
\arguments{
\item{IRsample}{cdr3aa column is needed, other(V,D,J,cdr3nt,type) are optional.}
}
\value{
saturation_matrix  18 rows and 8 columns
}
\description{
By 1  2  3  4  5  6  7  8  9 10 20 30 40 50 60 70 80 90 percent
To counts the reads/distinct cdr3aa number/QC/diversity
}
\examples{
#' plot_cdr3_stat(stat_matrix,'cdr3aa-distribution-stat.pdf')
}
\keyword{plot_cdr3_stat}
