% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_IR_basic.R
\name{stat_IR_basic}
\alias{stat_IR_basic}
\title{basic bstatistics for summary table from igblastn result}
\usage{
stat_IR_basic(IR_basic_df)
}
\value{
A list contain:v/d/j usage, top 10 cdr3,
}
\description{
analysis the summary table from igblastn result, v/d/j/cdr3aa/
THe other columns are optional: type,~~~
cdr3aa_diversity=cdr3aa_diversity,
v_usage=v_usage,
d_usage=d_usage,
j_usage=j_usage,
all_cdr3aa_usage=all_cdr3aa_usage,
cdr3aa_length_usage=cdr3aa_length_usage,
top10_cdr3aa=top10_cdr3aa,
cdr3aa_length_v=cdr3aa_length_v,
cdr3aa_stat=cdr3aa_stat,
vj_usage_matrix=vj_usage_matrix,
cdr3aa_length_cdr3aa=cdr3aa_length_cdr3aa
}
\examples{
#' stat_IR_basic(IR_basic_df)
}
\keyword{stat_IR_basic}
