% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_j_circle.R
\name{v_j_circle}
\alias{v_j_circle}
\title{Circle plot for V and J usage}
\usage{
v_j_circle(vj_usage_matrix, file_out)
}
\arguments{
\item{vj_usage_matrix}{columns and rows correspond to V and J segments respectively}

\item{file_out}{the filename of output figue,should end up with pdf or png}
}
\description{
It will generate a published quality figure according to the V and J usage marix.
}
\examples{
#' v_j_circle(vj_usage_matrix,'v-j-circle.pdf')
}
\keyword{circle}
