% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_IR_basic.R
\name{stat_IR_basic}
\alias{stat_IR_basic}
\title{basic bstatistics for summary table from igblastn result}
\usage{
stat_IR_basic(IR_basic_df)
}
\value{
A list contain:v/d/j usage, top 10 cdr3,
}
\description{
analysis the summary table from igblastn result, which must  includes the following fields:
Variable (V) segment name.
Diversity (D) segment name for some of the receptor chains (TRB, TRD and IGH). Set to . if not aplicable or D segment was not identified.
Joining (J) segment name.
Complementarity determining region 3 nucleotide sequence (CDR3nt). CDR3 starts with Variable region reference point (conserved Cys residue) and ends with Joining segment reference point (conserved PheTry).
Translated CDR3 sequence (CDR3aa).
THe other columns are optional: type,~~~
}
\examples{
#' stat_IR_basic(IR_basic_df)
}
\keyword{stat_IR_basic}
